requirements.txt
#Django
Django==3.2.7
python-decouple==3.4
python-dotenv==0.19.1

#Django Rest Framework
djangorestframework==3.12.4
django-filter==21.1
markdown==3.3.4
Pillow==9.0.0

#Documentation API
drf-spectacular==0.21.0

#Optional tools
pylint-django==2.4.4 